Changed multinomial from unsigned long long int, to unsigned long. This will probably mean overflow for large coefficients but there is no simple cure. This MUST go into the next CRAN submit

+++ src/multicool.cpp
@@ -301,7 +301,7 @@ NumericVector multinomCoeff(NumericVector x){
     v.at(i) = x[i];
   }
   
-  unsigned long long int u = multinomial::multi<unsigned long long int>(v);
+  unsigned long u = multinomial::multi<unsigned long>(v);
   
   NumericVector r = NumericVector::create(u);
